🎨 Fix coding style in Pencil Drawing

+++ src/App/ColorPicker/ColorPicker.cpp
@@ -1,9 +1,8 @@
-#include "ColorPicker.hpp"
+#include "App/ColorPicker/ColorPicker.hpp"
 #include "App/const.hpp"
 
 namespace MyGimp {
-ColorPicker::ColorPicker()
-{
+ColorPicker::ColorPicker() {
     if (!colorTexture.loadFromFile("Assets/ColorPicker.jpg"))
         return;
     colorSprite.setTexture(colorTexture);
@@ -17,20 +16,27 @@ void ColorPicker::draw(sf::RenderWindow &window) {
 sf::Color ColorPicker::handleInput(sf::Event &event) {
     if (event.type == sf::Event::MouseButtonPressed) {
         if (event.mouseButton.button == sf::Mouse::Left) {
-            // Use the event's window-relative coordinates instead of the global mouse position.
+            // Use the event's window-relative coordinates
             int mouseX = event.mouseButton.x;
             int mouseY = event.mouseButton.y;
             sf::Vector2f spritePos = colorSprite.getPosition();
             sf::Vector2u textureSize = colorTexture.getSize();
-            textureSize.x = static_cast<unsigned int>(textureSize.x * colorSprite.getScale().x);
-            textureSize.y = static_cast<unsigned int>(textureSize.y * colorSprite.getScale().y);
-            if (mouseX >= static_cast<int>(spritePos.x) && mouseX <= static_cast<int>(spritePos.x) + static_cast<int>(textureSize.x) &&
-                mouseY >= static_cast<int>(spritePos.y) && mouseY <= static_cast<int>(spritePos.y) + static_cast<int>(textureSize.y)) {
+            textureSize.x = static_cast<unsigned int>(textureSize.x *
+                colorSprite.getScale().x);
+            textureSize.y = static_cast<unsigned int>(textureSize.y *
+                colorSprite.getScale().y);
+            if (mouseX >= spritePos.x &&
+                mouseX <= spritePos.x + textureSize.x &&
+                mouseY >= spritePos.y &&
+                mouseY <= spritePos.y + textureSize.y) {
                 sf::Image img = colorTexture.copyToImage();
                 sf::Color pickedColor = img.getPixel(
-                    static_cast<unsigned int>((mouseX - static_cast<int>(spritePos.x)) / colorSprite.getScale().x),
-                    static_cast<unsigned int>((mouseY - static_cast<int>(spritePos.y)) / colorSprite.getScale().y)
-                );
+                    static_cast<unsigned int>((mouseX -
+                        static_cast<int>(spritePos.x))
+                        / colorSprite.getScale().x),
+                    static_cast<unsigned int>((mouseY -
+                        static_cast<int>(spritePos.y))
+                        / colorSprite.getScale().y));
                 return pickedColor;
             }
         }
@@ -38,14 +44,12 @@ sf::Color ColorPicker::handleInput(sf::Event &event) {
     return sf::Color::Transparent;
 }
 
-void ColorPicker::setPosition(float x, float y)
-{
+void ColorPicker::setPosition(float x, float y) {
     colorSprite.setPosition(x, y);
 }
 
-void ColorPicker::setScale(float scaleX, float scaleY)
-{
+void ColorPicker::setScale(float scaleX, float scaleY) {
     colorSprite.setScale(scaleX, scaleY);
 }
 
-} // namespace MyGimp
+}  // namespace MyGimp
